Add changevalue, hanldeZero, and handleChange functions to numerical buttons.
The following functions were created to help handle input:
changeValue, handleZero, handleChange.

changeValue - An encompassing function that calls hanzleZero or
              handleChange based on the value of the numerical button
              presed.

handleZero - Handles the change if the button pressed is a zero.

handleChange - Handles the change when the button pressed is not zero.

Created a global state object with the following properties:
total, memoryTotal, valueOne, valueTwo, operation, lastkeyPressed.