Refactor format.js for readability and functionality
Format.js is now the main logic behind handling what the calculator displays.

If the data provided is so large that an exponent is needed to represent it
in computers, the calculator throws an error.

If the data is too small, that it requires an exponent, the calculator truncates
it and shows only the first 16 relevant digits.

Format.js also contains the helper functions truncate and addCommas.